Reworked scoring system.
Now instead of generating all the possible
matches at the league start, it just tracks
your score against every other player in the
league, as you report matches.

After giving it some thought I realized that
players! was going to be much huger for leagues
than the old log(players) was for tournaments,
and things would probably break if we got above
like 7 or 8 players, which was clearly not ideal.